refactor: свернуть json_pipeline в single-source github_popular_pipeline
json_pipeline был generic-циклом по type:"json", но обслуживал ровно один
источник (github_new_popular). Дискриминатор «влезает в общий цикл» —
регулярность fetch'а, а не формат payload'а (даже второй JSON-источник,
steam, не влез из-за двухстадийного fetch). Снимаем ложную генерализацию:
явный single-source модуль в грейне steam_pipeline (_SOURCE_TYPE + тонкая
петля + per-source §IV-изоляция). Generic-раннер вернём, когда появятся
≥2 объединяемых источника.

- json_pipeline.py → github_popular_pipeline.py; run_json_pipeline → run_github_popular_pipeline
- sources.json type "json" → dedicated "github_popular"
- pipeline_config._SUPPORTED_TYPES: +github_popular, −json (мёртвое значение)
- .importlinter, test_import_contracts, test_package_importable: имя модуля
- workflow-шаг "Run JSON sources pipeline" → "Run GitHub popular pipeline"
- доки + стейл-комментарии

events_pipeline (тот же запах) — follow-up #276.


closes #275
