+++ tests/test_package_importable.py
@@ -24,9 +24,9 @@ _MODULES = [
     "events_pipeline",
     "gemini_enricher",
     "generic_pipeline",
+    "github_popular_pipeline",
     "github_trending_pipeline",
     "http_fetch",
-    "json_pipeline",
     "kinozal_auth",
     "kinozal_pipeline",
     "pipeline_config",
